ajax.php fix  - function already defined header fix 2
